# internal/service/resources/api/openapi.yaml
...
openapi: "3.0.0"
info:
  title: O2IMS Infrastructure Inventory API
  version: 1.0.0
  description: |
    API for O2IMS Infrastructure Inventory
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
- url: http://localhost:8000
...
paths:
  /o2ims-infrastructureInventory/api_versions:
    get:
      operationId: getAllVersions
      summary: Get API versions
      description: |
        Returns the complete list of API versions implemented by the service.
      tags:
      - metadata
      responses:
        '200':
          description: |
            Successfully obtained the complete list of versions.
          content:
            application/json:
              schema:
                $ref: "../../common/api/openapi.yaml#/components/schemas/APIVersions"
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
  /o2ims-infrastructureInventory/v1:
    get:
      operationId: getCloudInfo
      summary: Get O-Cloud info
      description: |
        Returns the details of the O-Cloud instance.
      parameters:
      - $ref: "../../common/api/openapi.yaml#/components/parameters/allFields"
      - $ref: "../../common/api/openapi.yaml#/components/parameters/excludeFields"
      - $ref: "../../common/api/openapi.yaml#/components/parameters/fields"
      tags:
      - metadata
      responses:
        '200':
          description: |
            Successfully obtained the details of the O-Cloud instance.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OCloudInfo"
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
  /o2ims-infrastructureInventory/v1/api_versions:
    get:
      operationId: getMinorVersions
      summary: Get minor API versions
      description: |
        Returns the list of minor API versions implemented for this major version of the API.
      tags:
      - metadata
      responses:
        '200':
          description: |
            Success
          content:
            application/json:
              schema:
                $ref: "../../common/api/openapi.yaml#/components/schemas/APIVersions"
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
  /o2ims-infrastructureInventory/v1/deploymentManagers:
    get:
      operationId: getDeploymentManagers
      summary: Get deployment managers
      description: |
        Returns the list of deployment managers.
      parameters:
      - $ref: "../../common/api/openapi.yaml#/components/parameters/allFields"
      - $ref: "../../common/api/openapi.yaml#/components/parameters/excludeFields"
      - $ref: "../../common/api/openapi.yaml#/components/parameters/fields"
      - $ref: "../../common/api/openapi.yaml#/components/parameters/filter"
      tags:
      - deploymentManagers
      responses:
        '200':
          description: |
            Successfully obtained the list of deployment managers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeploymentManager'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
  /o2ims-infrastructureInventory/v1/deploymentManagers/{deploymentManagerId}:
    get:
      operationId: getDeploymentManager
      summary: Get deployment manager
      description: |
        Returns the details of a deployment manager.
      parameters:
      - $ref: "#/components/parameters/deploymentManagerId"
      tags:
      - deploymentManagers
      responses:
        '200':
          description: |
            Successfully obtained the details of the deployment manager.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentManager"
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '404':
          description: The specified entity was not found.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
  /o2ims-infrastructureInventory/v1/subscriptions:
    get:
      operationId: getSubscriptions
      summary: Get subscriptions
      description: |
        Returns the list of inventory subscriptions.
      parameters:
      - $ref: "../../common/api/openapi.yaml#/components/parameters/excludeFields"
      - $ref: "../../common/api/openapi.yaml#/components/parameters/allFields"
      - $ref: "../../common/api/openapi.yaml#/components/parameters/fields"
      - $ref: "../../common/api/openapi.yaml#/components/parameters/filter"
      tags:
      - subscriptions
      responses:
        '200':
          description: |
            Successfully obtained the list of subscriptions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscription'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
    post:
      operationId: createSubscription
      summary: Create subscriptions
      description: |
        Creates an inventory subscriptions.
      tags:
      - subscriptions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Subscription'
      responses:
        '201':
          description: |
            Successfully created the subscriptions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
  /o2ims-infrastructureInventory/v1/subscriptions/{subscriptionId}:
    get:
      operationId: getSubscription
      summary: Get subscription
      description: |
        Returns the details of a subscription.
      parameters:
      - $ref: "#/components/parameters/subscriptionId"
      tags:
      - subscriptions
      responses:
        '200':
          description: |
            Successfully obtained the details of the subscription.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '404':
          description: The specified entity was not found.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
    delete:
      operationId: deleteSubscription
      summary: Delete subscription
      description: |
        Deletes a subscription.
      parameters:
      - $ref: "#/components/parameters/subscriptionId"
      tags:
      - subscriptions
      responses:
        '200':
          description: |
            Successfully deleted the subscription.
        '404':
          description: The specified entity was not found.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
  /o2ims-infrastructureInventory/v1/resourceTypes:
    get:
      operationId: getResourceTypes
      summary: Get resource types
      description: |
        Returns the list of resource types.
      parameters:
      - $ref: "../../common/api/openapi.yaml#/components/parameters/allFields"
      - $ref: "../../common/api/openapi.yaml#/components/parameters/excludeFields"
      - $ref: "../../common/api/openapi.yaml#/components/parameters/fields"
      - $ref: "../../common/api/openapi.yaml#/components/parameters/filter"
      tags:
      - resourceTypes
      responses:
        '200':
          description: |
            Successfully obtained the list of resource types.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResourceType"
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
  /o2ims-infrastructureInventory/v1/resourceTypes/{resourceTypeId}:
    get:
      operationId: getResourceType
      summary: Get a resource type
      description: |
        Returns the details of a resource type.
      parameters:
      - $ref: "#/components/parameters/resourceTypeId"
      tags:
      - resourceTypes
      responses:
        '200':
          description: |
            Successfully obtained the details of the resource type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceType"
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '404':
          description: The specified entity was not found.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
  /o2ims-infrastructureInventory/v1/resourcePools:
    get:
      operationId: getResourcePools
      summary: Get resource pools
      description: |
        Returns the list of resource pools.
      parameters:
      - $ref: "../../common/api/openapi.yaml#/components/parameters/allFields"
      - $ref: "../../common/api/openapi.yaml#/components/parameters/excludeFields"
      - $ref: "../../common/api/openapi.yaml#/components/parameters/fields"
      - $ref: "../../common/api/openapi.yaml#/components/parameters/filter"
      tags:
      - resourcePools
      responses:
        '200':
          description: |
            Successfully obtained the list of resource pools.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResourcePool"
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
...
components:
  parameters:
    deploymentManagerId:
      name: deploymentManagerId
      description: |
        Unique identifier of a deployment manager.
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 65221564-8b05-416a-bfc3-c250bc64e1aa
    subscriptionId:
      name: subscriptionId
      description: |
        Unique identifier of a subscription.
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: aa83d0ba-a5ee-4f1f-be84-f334b21166cf
...
    resourceTypeId:
      name: resourceTypeId
      description: |
        Unique identifier of a resource type.
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 31a69575-3f90-4eb5-b8fa-8f8e23f99cf0
  schemas:
...
    OCloudInfo:
      type: object
      properties:
        oCloudId:
          type: string
          format: uuid
          description: Identifier of the O-Cloud instance. Internally generated within an O-Cloud instance.
          example: "262c8f17-52b5-4614-9e56-812ae21fa8a7"
        globalcloudId:
          type: string
          format: uuid
          description: | 
            Identifier of the O-Cloud instance assigned by the SMO. This identifier is globally unique across O-Cloud 
            instances known to the SMO. This value was provided by the SMO at cloud genesis and is stored in the O-Cloud
            IMS Inventory.
          example: "8c1c151d-2899-4c96-b76f-4fe92064b57b"
        name:
          type: string
          description: Human readable name of the O-Cloud as identified by the SMO at cloud genesis.
          example: "my-cloud"
        description:
          type: string
          description: Human readable description of the O-Cloud as provided by the SMO at cloud genesis.
          example: "My cloud"
        serviceUri:
          type: string
          description: The URI root to all services provided by the O2ims interface. Inventory is one of these services.
          example: "http://localhost:8000"
        extensions:
          type: object
          example:
            hub: "hub0"
            country: "ES"
      required:
      - oCloudId
      - globalcloudId
      - name
      - description
      - serviceUri
...